Add simplified Dockerfiles for running TF
These Dockerfiles are intended to match exactly the feature set of the
Docker images currently available on Docker Hub for
tensorflow/tensorflow. Due to security concerns we're going to have to
remove those containers from the TensorFlow source tree itself -- which
makes some sense, as the TensorFlow team does not make use of them, and
so they do not need to be subject to the same rigorous standards as
other containers which will remain in the TF repository.

Furthermore, due to maintenance concerns, the complex "partials" system (which I
originally designed) has not been working out for us, and we would like
to deprecate it in favor of only a minimal set of containers.