install and run jupyter (with tensorflow) as non-root user
The entire installation of jupyter and tensorflow is done as
${TENSORFLOW_USER} with pip. Folder /home/${TENSORFLOW_USER}/.local/bin
is added to the environment variable PATH. A notebook folder
/home/${TENSORFLOW_USER}/notebooks is created and the tutorials
are install in this folder.

Finally the command

	source /etc/bash.bashrc && jupyter notebook --notebook-dir=${TENSORFLOW_NOTEBOOK_DIR} --ip 0.0.0.0 --no-browser --allow-root

is run as TENSORFLOW_USER.

The only thing still bothering me is the call to ldconfig in /etc/bash.bashrc.
I don't understand why we need this. And it can't be done as non-root user.